Add Functions to Fetch Popular People and Their Details
Added two new functions: fetchPeople to get a list of popular actors and celebrities, and fetchPersonDetails to get detailed information about a specific person. Both functions connect to the Movie Database API and handle errors if the request fails. This will help us show popular people and their profiles in the app.