Cleaned Up and Renamed Interfaces for Actors and Movies
Reorganized and renamed several parts of the code to avoid duplicates and make things clearer. "People" interfaces were changed to "Actors" to better show that they refer to movie actors. Also removed or fixed repeated and outdated sections, like multiple Review, MovieDetailsProps, and KnownFor definitions. This helps keep the code neat and easier to work with.