Added New Page to Show List of Popular Actors
Created a new page that displays a list of popular actors using data from the movie database. It includes a loading spinner, error message handling, and pagination so users can move through different pages of actors. Each actor also has a button to add them to favourites. This makes it easier for users to browse and save their favorite stars.